move objects to use ValueObject namespace
