+++ utils/RectorGenerator/src/Command/CreateRectorCommand.php
@@ -5,10 +5,10 @@ namespace Rector\Utils\RectorGenerator\Command;
 use Nette\Utils\FileSystem;
 use Nette\Utils\Json;
 use Nette\Utils\Strings;
-use Rector\Utils\RectorGenerator\Configuration\Configuration;
 use Rector\Utils\RectorGenerator\Configuration\ConfigurationFactory;
 use Rector\Utils\RectorGenerator\Contract\ContributorCommandInterface;
 use Rector\Utils\RectorGenerator\TemplateVariablesFactory;
+use Rector\Utils\RectorGenerator\ValueObject\Configuration;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
